refactor: move mapper test to correctly dir
