+++ test/core/application/@mocks/fake-map-stub.ts
